requirements.txt
beautifulsoup4==4.12.2
ffmpeg_python==0.2.0
python_magic==0.4.27
soundfile==0.12.1
vosk==0.3.45
